extracted github entities into classes
